+++ packages/backend/src/models/DashboardModel/DashboardModel.mock.ts
@@ -1,5 +1,6 @@
 import { Ability } from '@casl/ability';
 import {
+    ChartKind,
     CreateDashboard,
     CreateDashboardChartTile,
     Dashboard,
@@ -33,7 +34,6 @@ const tileWithoutId: CreateDashboardChartTile = {
     w: 200,
     properties: {
         savedChartUuid: '123',
-        title: 'title 123',
     },
 };
 const tileWithId: DashboardChartTile = {
@@ -96,7 +96,6 @@ export const addDashboardVersionWithoutChart: DashboardVersionedFields = {
             ...tileWithoutId,
             properties: {
                 savedChartUuid: null,
-                title: null,
             },
         },
     ],
@@ -132,6 +131,9 @@ export const savedChartEntry: SavedChartTable['base'] = {
     name: 'chart name',
     description: 'My description',
     created_at: new Date(),
+    last_version_chart_kind: ChartKind.VERTICAL_BAR,
+    last_version_updated_at: new Date(),
+    last_version_updated_by_user_uuid: undefined,
 };
 
 export const dashboardEntry: DashboardTable['base'] = {