+++ packages/frontend/src/components/DashboardTiles/TileBase/index.tsx
@@ -11,6 +11,7 @@ import { Tooltip } from '@mantine/core';
 import { useHover, useToggle } from '@mantine/hooks';
 import React, { ReactNode, useState } from 'react';
 import ChartUpdateModal from '../TileForms/ChartUpdateModal';
+import TileUpdateChartTitle from '../TileForms/TileUpdateChartTitle';
 import TileUpdateModal from '../TileForms/TileUpdateModal';
 import {
     ButtonsWrapper,
@@ -24,6 +25,7 @@ import {
 type Props<T> = {
     isEditMode: boolean;
     title: string;
+    chartName?: string;
     titleHref?: string;
     description?: string;
     tile: T;
@@ -38,6 +40,7 @@ type Props<T> = {
 const TileBase = <T extends Dashboard['tiles'][number]>({
     isEditMode,
     title,
+    chartName,
     description,
     tile,
     isLoading,
@@ -48,7 +51,9 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
     extraHeaderElement,
     titleHref,
 }: Props<T>) => {
-    const [isEditing, setIsEditing] = useState(false);
+    const [isReplacingChart, setIsReplacingChart] = useState(false);
+    const [isEditingTitle, setIsEditingTitle] = useState(false);
+
     const [isHovering, setIsHovering] = useState(false);
     const { hovered: containerHovered, ref: containerRef } = useHover();
     const { hovered: titleHovered, ref: titleRef } =
@@ -116,9 +121,18 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
                                                     <>
                                                         <MenuItem2
                                                             icon="edit"
-                                                            text="Edit tile content"
+                                                            text="Edit title"
                                                             onClick={() =>
-                                                                setIsEditing(
+                                                                setIsEditingTitle(
+                                                                    true,
+                                                                )
+                                                            }
+                                                        />
+                                                        <MenuItem2
+                                                            icon="exchange"
+                                                            text="Replace chart in tile"
+                                                            onClick={() =>
+                                                                setIsReplacingChart(
                                                                     true,
                                                                 )
                                                             }
@@ -182,26 +196,42 @@ const TileBase = <T extends Dashboard['tiles'][number]>({
                     {tile.type === DashboardTileTypes.SAVED_CHART ? (
                         <ChartUpdateModal
                             className="non-draggable"
-                            opened={isEditing}
+                            opened={isReplacingChart}
                             tile={tile}
-                            onClose={() => setIsEditing(false)}
-                            onConfirm={(chartTile) => {
-                                onEdit(chartTile);
-                                setIsEditing(false);
+                            onClose={() => setIsReplacingChart(false)}
+                            onConfirm={(newChartTile) => {
+                                onEdit(newChartTile);
+                                setIsReplacingChart(false);
                             }}
                         />
                     ) : (
                         <TileUpdateModal
                             className="non-draggable"
-                            isOpen={isEditing}
+                            isOpen={isReplacingChart}
                             tile={tile}
-                            onClose={() => setIsEditing(false)}
-                            onConfirm={(data) => {
-                                onEdit(data);
-                                setIsEditing(false);
+                            onClose={() => setIsReplacingChart(false)}
+                            onConfirm={(newTile) => {
+                                onEdit(newTile);
+                                setIsReplacingChart(false);
                             }}
                         />
                     )}
+                    <TileUpdateChartTitle
+                        isOpen={isEditingTitle}
+                        placeholder={chartName || ''}
+                        title={title}
+                        onClose={() => setIsEditingTitle(false)}
+                        onConfirm={(newTitle) => {
+                            onEdit({
+                                ...tile,
+                                properties: {
+                                    ...tile.properties,
+                                    title: newTitle,
+                                },
+                            });
+                            setIsEditingTitle(false);
+                        }}
+                    />
                 </>
             )}
         </TileBaseWrapper>